+++ packages/jest-cli/src/init/__tests__/init.test.js
@@ -25,17 +25,13 @@ const consoleLog = console.log;
 
 describe('init', () => {
   beforeEach(() => {
-    // $FlowFixMe mock console.log to reduce noise from the tests
     console.log = jest.fn();
-    // $FlowFixMe mock
     fs.writeFileSync = jest.fn();
   });
 
   afterEach(() => {
     jest.clearAllMocks();
-    // $FlowFixMe
     console.log = consoleLog;
-    // $FlowFixMe mock
     fs.writeFileSync = writeFileSync;
   });